tentando concertar após int64

+++ models.go
@@ -14,7 +14,7 @@ type Reading struct {
 	Timestamp string  `json:"timestamp"`
 	Value     float64 `json:"value"`
 	TorqueValues  []float64 `json:"torque_values"`
-    AsmTimes      []int    `json:"asm_times"`
-    MotionWastes  []int    `json:"motion_wastes"`
+    AsmTimes      []int64    `json:"asm_times"`
+    MotionWastes  []int64    `json:"motion_wastes"`
     SetValue      float64   `json:"set_value"`
 }
